add positioning image bgs, fix positioning bug
basically when you selected a position, when it tried to generate the script it didnt divide by two, meaning the
position was 2x the place it was supposed to be.

you can now drag and drop images into the positioning window to have a custom background. keep in mind that this
is a WIP and im only committing this now to fix the bug above